Removed DB constraints on Item_locations
This was causing unexpected behavior when both item_id's exxisted in the found_item and lost_item table. I reviewed each form handler and since it was impossible to store a location in the table that did not have an id from either of these tables, and since we specify whether it is lost or found in item_type. I decieded to remove the constraint to fix the issue